refactor: add useCases folder structure to components
