+++ src/components/feature/calculadora13/FormCalculadora13.tsx
@@ -4,7 +4,7 @@ import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
 import { Select } from "@/components/ui/select";
 import { useForm, SubmitHandler } from "react-hook-form";
-import { calculateDecimoTerceiro } from "./utils/calculateDecimoTerceiro";
+import { calculateDecimoTerceiro } from "./useCases/calculateDecimoTerceiro";
 import Markdown from "react-markdown";
 import { text } from "./utils/text";
 import { useState } from "react";
